Inline the IsSelected* and NotTransparent* methods
Profiling indicates that the IsSelected* and NotTransparent* methods
cause a number of unnecessary bounds checks by accessing each pixel
using the GetPoint method in the MaskSurface and Surface classes.

Removed IsSelectedTarget, IsSelectedSource, NotTransparentTarget and
NotTransparentSource methods.
Updated the ClippedOrMaskedSource method to use the MaskSurface
GetPointUnchecked method.
Updated the PrepareTargetPoints and PrepareSourcePoints methods to
use pointers when reading the images and masks.