Report rendering errors to the effect dialog
This prevents multiple message boxes from being displayed if an error
occurs when rendering and the user tries to exit the effect dialog.
The effect dialog will now show a single message box when the user
changes the sample area size.

Linking the error dialog boxes to the sample area size allows for the
case that only some of the sample area contains transparent pixels.
Although mixing transparent and non-transparent pixels can produce
artifacts because the Resynthesizer algorithm was not designed to
process transparent pixels.