Allow selections that extend to the image edges
The code that detects if the whole image is selected would produce a
false-positive when only the edges of a image were selected.
The new code checks if there are any unselected areas in the image,
and will only show an error if that is not the case.

Added a shortcut that skips further selection checks when the
configuration dialog is open.