Make the automatic rendering optional
By default the configuration dialog will automatically start rendering
when the user changes a value.
This may be undesirable for large images because the algorithm can take
a significant amount of time to process the image whenever a setting is
changed.

The configuration dialog now has a checkbox that allows the user to
disable the automatic rendering for the current session.