Improve responsiveness when canceling the effect
Replaced the abort callback delegate with a CancellationToken.
The ContentAwareFill method now signals cancellation by throwing an
OperationCanceledException.
The behavior of the ResynthesizerRunner Run method returning null on
cancallation is unchanged.

Cancellation checks are now performed at the start of the main loops in
the Synthesize and TryPoint methods instead of only when reporting
progress.

Fixes #5