Remove the ResynthesizerParameters class
This class is unnecessary because all but one of the properties are
constants.
The constants that are used by the algorithm have been moved into
Resynthesizer.cs.

The TileHorizontal and TileVertical properties that were used in
WrapOrClip were always false, so the method has been simplified to
reflect that behavior and renamed to InTargetBounds.