Make the Synthesize function multi-threaded
It will now use all logical CPU cores when processing the image data.

This massively improves the performance of the rendering process. On a
test image with a 4593x4719 pixel selection (25,921,467 pixels square),
the rendering time went from one minute fourteen seconds with a single
thread down to 5 seconds with 32 threads.