Optimize the algorithm in PrepareSortedOffsets
The new algorithm was created by Rick Brewster on the Paint.NET
Discord server:

https://discord.com/channels/751565809972936735/1010984760782377160/1303378279155372072

It is normally 20+ seconds faster than the previous code which sorted
the items with MemoryExtensions.Sort after adding them to the list.
The two implementations place the items in a different order, but that
doesn't matter as the algorithm will scan the entire list in
PrepareNeighbors.