Add a ResynthesizerRandom class
This class is a thin wrapper that allows the underlying PRNG
implementation to be changed without needing to update the
calling code.

The current PRNG is System.Random, but it is worth experimenting with
other implementations due to how frequently the Resynthesizer algorithm
uses pseudo-random numbers.
System.Random.Next(int maxValue) is one of the most called methods in
the algorithm's inner loops, my current performance profiles put it at
~12% to ~15% of the total CPU time.